Update SpaceInvadersCollection2.py
A hopeful attempt at shooting detection for the aliens below each row. This may not work or cause major lag in the game, if so revert back to a previous build in order to fix it as it is very experimental and I don't know the result of the code.